requirements.txt
dlib>=19.3.0
numpy
Pillow
face_recognition
kafka-python
sklearn
pandas
opencv-python
boto3
